Support multi-stage stream processing by consumers returning effects
